Removed gsap animations on mount and unmount
The gsap animations which were initially set to run on mounting and unmounting of components caused build failures on Netlify. By removing the calls to these animations in various Vue.js components, the issue is expected to be resolved. This change will effectively prevent unnecessary execution of the animations during the build process.